// my-cooking-app/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          bio: string | null
          created_at: string
          favourite_cuisine: Database["public"]["Enums"]["cuisine_type"] | null
          id: string
          number_of_uploaded_recipes: number
          profile_picture_url: string | null
          updated_at: string
          username: string
        }
        Insert: {
          bio?: string | null
          created_at?: string
          favourite_cuisine?: Database["public"]["Enums"]["cuisine_type"] | null
          id: string
          number_of_uploaded_recipes?: number
          profile_picture_url?: string | null
          updated_at?: string
          username: string
        }
        Update: {
          bio?: string | null
          created_at?: string
          favourite_cuisine?: Database["public"]["Enums"]["cuisine_type"] | null
          id?: string
          number_of_uploaded_recipes?: number
          profile_picture_url?: string | null
          updated_at?: string
          username?: string
        }
        Relationships: []
      }
// ...
      recipes: {
        Row: {
          author_id: string
          created_at: string
          cuisine: Database["public"]["Enums"]["cuisine_type"] | null
          description: string | null
          id: string
          image_url: string | null
          ingredients: string[]
          instructions: string[]
          number_of_likes: number
          number_of_saves: number
          recipe_type: Database["public"]["Enums"]["recipe_kind"] | null
          title: string
        }
        Insert: {
          author_id: string
          created_at?: string
          cuisine?: Database["public"]["Enums"]["cuisine_type"] | null
          description?: string | null
          id?: string
          image_url?: string | null
          ingredients: string[]
          instructions: string[]
          number_of_likes?: number
          number_of_saves?: number
          recipe_type?: Database["public"]["Enums"]["recipe_kind"] | null
          title: string
        }
        Update: {
          author_id?: string
          created_at?: string
          cuisine?: Database["public"]["Enums"]["cuisine_type"] | null
          description?: string | null
          id?: string
          image_url?: string | null
          ingredients?: string[]
          instructions?: string[]
          number_of_likes?: number
          number_of_saves?: number
          recipe_type?: Database["public"]["Enums"]["recipe_kind"] | null
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "recipes_author_id_fkey"
            columns: ["author_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never